rename components to views
